+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 
+### 0.0.2
+- Added `GeoStatsEvaluator`, `ScoreReporter` and `MetricsHelper` class.
+- Updated `example.py` file
+- Updated README.md file with usage examples
+- Added unit tests for new classes
+
+
 ### 0.0.1
 - Introduces `pathways-bench` to generate Tessellating Intersection Polygons.
 - Added unit test cases